修改工作流，修改配置文件

+++ libwrt.sh
@@ -1,3 +1,34 @@
+#!/usr/bin/env bash
+set -Eeuo pipefail
+
+# 修改默认IP & 固件名称 & 编译署名和时间
+sed -i 's/192.168.1.1/192.168.1.254/g' package/base-files/files/bin/config_generate
+sed -i "s/hostname='.*'/hostname='Roc'/g" package/base-files/files/bin/config_generate
+luci_system_js="feeds/luci/modules/luci-mod-status/htdocs/luci-static/resources/view/status/include/10_system.js"
+firmware_version_anchor="_('Firmware Version'), (L.isObject(boardinfo.release) ? boardinfo.release.description + ' / ' : '') + (luciversion || ''),"
+grep -Fq "$firmware_version_anchor" "$luci_system_js" || { echo "Error: LuCI firmware version anchor was not found in $luci_system_js" >&2; exit 1; }
+sed -i "s#_('Firmware Version'), (L\.isObject(boardinfo\.release) ? boardinfo\.release\.description + ' / ' : '') + (luciversion || ''),# \
+            _('Firmware Version'),\n \
+            E('span', {}, [\n \
+                (L.isObject(boardinfo.release)\n \
+                ? boardinfo.release.description + ' / '\n \
+                : '') + (luciversion || '') + ' / ',\n \
+            E('a', {\n \
+                href: 'https://github.com/chicha9009/openwrt-ci/releases',\n \
+                target: '_blank',\n \
+                rel: 'noopener noreferrer'\n \
+                }, [ 'Built by 不如吃茶去 $(date "+%Y-%m-%d %H:%M:%S")' ])\n \
+            ]),#" "$luci_system_js"
+
+
+
 rm -rf package/emortal/luci-app-athena-led
 git clone --depth=1 https://github.com/NONGFAH/luci-app-athena-led package/luci-app-athena-led
 chmod +x package/luci-app-athena-led/root/etc/init.d/athena_led package/luci-app-athena-led/root/usr/sbin/athena-led
+
+rm -rf feeds/luci/applications/luci-app-eqos
+
+git clone --depth=1 https://github.com/chicha9009/luci-app-eqos.git feeds/luci/applications/luci-app-eqos
+
+./scripts/feeds update  -a
+./scripts/feeds install -a
